+++ sdmx-dl-provider-ri/src/main/java/internal/sdmxdl/ri/web/DotStatRestQueries.java
@@ -1,6 +1,7 @@
 package internal.sdmxdl.ri.web;
 
 import internal.util.rest.RestQueryBuilder;
+import org.checkerframework.checker.nullness.qual.NonNull;
 import sdmxdl.DataFilter;
 import sdmxdl.DataStructureRef;
 import sdmxdl.DataflowRef;
@@ -42,6 +43,16 @@ public class DotStatRestQueries implements RiRestQueries {
                 .path(flowRef.getId())
                 .path(key.toString())
                 .param("format", "compact_v2");
+    }
+
+    @Override
+    public DataStructureRef peekStructureRef(DataflowRef ref) {
+        return getStructureRefFromFlowRef(ref);
+    }
+
+    @NonNull
+    public static DataStructureRef getStructureRefFromFlowRef(@NonNull DataflowRef o) {
+        return DataStructureRef.of(o.getAgency(), o.getId(), o.getVersion());
     }
 
     public static final String DATASTRUCTURE_RESOURCE = "GetDataStructure";